pyproject.toml
[tool.poetry]
name = "eden"
version = "1.0.0"
description = ""
authors = ["Sam Craig <sam@growthat.agency>"]

[tool.poetry.dependencies]
python = "^3.9"
invoke = "^1.5.0"
isort = "^5.10.1"

[tool.poetry.dev-dependencies]
black = "^21.8b0"
jsonmerge = "^1.8.0"
PyYAML = "^5.4.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 80
